requirements.txt
requests~=2.31.0
flask~=3.0.2
gunicorn==21.2.0
psycopg2-binary==2.9.6
Pillow==10.2.0